requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.40.0
httpx>=0.27.0
anthropic>=0.34.0
